+++ ush/rtofs_ncoda_goes_qc.sh
@@ -159,7 +159,7 @@ ln -s $OCN_DATA_DIR/incoming/goes.b.$cut_dtg $OCN_DATA_DIR/incoming/goes.b
 $EXECrtofs/rtofs_ncoda_qc $cut_dtg goes > goes_qc.$cut_dtg.out
 err=$?; export err ; err_chk
 echo " error from rtofs_ncoda_qc=",$err
-mv fort.44 goes_qc.$cut_dtg.rej
+[[ -f fort.44 ]] && mv fort.44 goes_qc.$cut_dtg.rej
 
 #   cleanup
 #rm -f g16_*.*
