+++ ush/rtofs_ncoda_himawari_qc.sh
@@ -157,7 +157,7 @@ ln -s $OCN_DATA_DIR/incoming/himawari.b.$cut_dtg $OCN_DATA_DIR/incoming/himawari
 $EXECrtofs/rtofs_ncoda_qc $cut_dtg himawari > himawari_qc.$cut_dtg.out
 err=$?; export err ; err_chk
 echo " error from rtofs_ncoda_qc=",$err
-mv fort.44 himawari_qc.$cut_dtg.rej
+[[ -f fort.44. ]] && mv fort.44 himawari_qc.$cut_dtg.rej
 
 #   cleanup
