+++ ush/rtofs_ncoda_ssh_qc.sh
@@ -120,7 +120,7 @@ ln -s $OCN_DATA_DIR/incoming/ssh.b.$cut_dtg $OCN_DATA_DIR/incoming/ssh.b
 $EXECrtofs/rtofs_ncoda_qc $cut_dtg ssh > ssh_qc.$cut_dtg.out
 err=$?; export err ; err_chk
 echo " error from rtofs_ncoda_qc=",$err
-mv fort.44 ssh_qc.$cut_dtg.rej
+[[ -f fort.44 ]] && mv fort.44 ssh_qc.$cut_dtg.rej
 
 echo "*** Finished script $0 on hostname "$(hostname)' at time '$(date)
